commenting these out is going to make it worse not better.
the idea with rel2absolute is to turn "filename" into
"/home/dstoll/filename" so we can find it after we change
directories.

Try it by hand a sec...

    perl -MFile::Spec -e 'print "", File::Spec->rel2abs("supz.txt"), "\n"'

The problem with the daemonize is that we change our directory to
/ after the fork() so the "name.mpg" isn't going in the directory
where you're running the program, it's going in / ... which was
causing you permissions problems.

The correct solution is to work with the absolute paths and fix
those problems.  Everything else is just a temporary work around.